requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
twilio==8.12.0
geocoder==1.38.1
python-dateutil==2.8.2
pytz==2024.1 